Improve: Eliminate abstract requirements from NullAuthenticatable
