+++ src/NullAuthenticatable.php
@@ -4,57 +4,6 @@ namespace Mpyw\NullAuth;
 
 trait NullAuthenticatable
 {
-    /**
-     * Get the name of the unique identifier for the user.
-     *
-     * @return string
-     */
-    abstract public function getAuthIdentifierName();
-
-    /**
-     * Get the unique identifier for the user.
-     *
-     * @return mixed
-     */
-    abstract public function getAuthIdentifier();
-
-    /**
-     * Get the password for the user.
-     *
-     * @return string
-     */
-    public function getAuthPassword()
-    {
-        return '';
-    }
-
-    /**
-     * Get the token value for the "remember me" session.
-     *
-     * @return string
-     */
-    public function getRememberToken()
-    {
-        return '';
-    }
-
-    /**
-     * Set the token value for the "remember me" session.
-     *
-     * @param  string $value
-     * @return void
-     */
-    public function setRememberToken($value)
-    {
-    }
-
-    /**
-     * Get the column name for the "remember me" token.
-     *
-     * @return string
-     */
-    public function getRememberTokenName()
-    {
-        return '';
-    }
+    use GenericNullAuthenticatable,
+        Concerns\HasEloquentIdentifier;
 }
